Converted images again to webp and fixed all possible promises issues, page will always load regardless  what pokemon 1 -151 is.
I kept having uncaught in promise issues, suddenly realized I didn't have a catch statement to prrevent the uncaught problems that have arised.

Then I realized it wasn't a promise issue but the code I've writtento print the pokeon weaknesses. The goal was to compare the two arrays, i initially used two for loops and when a match was found I would decrease the counter. That's why and where the undefined problem came. Fxed it with a solution on stackexchange by combing filter() with every() to solve this problem