Attempting to refactor the code using separate functions and trying to make it cleaner overall
Code looked awful but ran. Was also far too small, over 1 second to fetch a pokemon is too slow and it looked ugly in between fetching, which is why i added the loader.

We are no longer just awaiting each statement, now that I have a better grasp of promises and async/await I'll be using them more often when necessary.